# docs/openapi/categories.yaml
openapi: 3.0.3
info:
  title: RVP Categories and Mirror API
  version: 0.1.0
servers:
  - url: https://visibleshelfstaging.up.railway.app
paths:
...
  /api/categories/mirror:
    post:
      summary: Trigger category mirror job (flag-gated)
      description: Returns 202 with jobId. If dryRun=true, perform a dry run.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [strategy]
              properties:
                strategy:
                  type: string
                  enum: [platform_to_gbp, gbp_to_platform]
                tenantId:
                  type: string
                dryRun:
                  type: boolean
                  default: true
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  accepted:
                    type: boolean
                  jobId:
                    type: string
                  strategy:
                    type: string
                  tenantId:
                    type: string
        '400':
          description: Bad Request (invalid strategy)
        '409':
          description: Feature disabled (FF_CATEGORY_MIRRORING off)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT